dump

+++ srcs/ft_handler_mouse_motion_julia.c
@@ -6,7 +6,7 @@
 /*   By: sclolus <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/08 17:12:40 by sclolus           #+#    #+#             */
-/*   Updated: 2017/07/15 01:19:38 by sclolus          ###   ########.fr       */
+/*   Updated: 2017/07/15 02:09:29 by sclolus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,6 @@
 
 void	ft_handler_mouse_motion_julia(int x, int y, void *param)
 {
-	ft_get_t_fractal_data()[JULIA].z0 = (t_complexe){(float)x / WINDOW_WIDTH - 1, (float)y / WINDOW_HEIGHT - 1};
+	ft_get_t_fractal_data()[JULIA].z0 = (t_complexe){(double)x / WINDOW_WIDTH - 1, (double)y / WINDOW_HEIGHT - 1};
 	ft_draw_fractal((t_mlx_data*)((t_mlx_data**)param)[0], JULIA);
 }
